refactor: move HNS cleanup to a higher level of abstraction
