+++ cns/restserver/ipam.go
@@ -12,7 +12,6 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/Azure/azure-container-networking/aitelemetry"
 	"github.com/Azure/azure-container-networking/cns"
 	"github.com/Azure/azure-container-networking/cns/filter"
 	"github.com/Azure/azure-container-networking/cns/logger"
@@ -1485,91 +1484,3 @@ func (service *HTTPRestService) GetIPFamilyCount() int {
 	return len(service.getIPFamiliesMap())
 }
 
-// normalizeMAC lowercases and converts hyphen-delimited MACs to colon-delimited.
-func normalizeMAC(mac string) string {
-	return strings.ToLower(strings.ReplaceAll(mac, "-", ":"))
-}
-
-// findEndpointStateByMAC returns the first EndpointState entry with a delegated NIC matching the given MAC.
-// A MAC corresponds to a single physical NIC, so at most one endpoint can match.
-// Caller must hold the service lock.
-func (service *HTTPRestService) findEndpointStateByMAC(mac string) (string, *IPInfo) {
-	for containerID, epInfo := range service.EndpointState {
-		for _, ipInfo := range epInfo.IfnameToIPMap {
-			if (ipInfo.NICType == cns.DelegatedVMNIC || ipInfo.NICType == cns.NodeNetworkInterfaceFrontendNIC) &&
-				normalizeMAC(ipInfo.MacAddress) == normalizeMAC(mac) {
-				return containerID, ipInfo
-			}
-		}
-	}
-	return "", nil
-}
-
-// ncAlreadyExistsForMAC returns true if the NC already exists in NC state with the given MAC.
-// Caller must hold the service lock.
-func (service *HTTPRestService) ncAlreadyExistsForMAC(ncID, mac string) bool {
-	nc, ok := service.state.ContainerStatus[ncID]
-	if !ok {
-		return false
-	}
-	return normalizeMAC(nc.CreateNetworkContainerRequest.NetworkInterfaceInfo.MACAddress) == normalizeMAC(mac)
-}
-
-// cleanupStaleHNSResources removes HNS endpoints/networks left behind by a previous NC
-// that used the same delegated NIC MAC and was unable to be cleaned up (eg. no CNI DEL was called), then deletes the stale EndpointState entries.
-func (service *HTTPRestService) cleanupStaleHNSResources(ncID, mac string) (returnErr error) {
-	service.Lock()
-	defer service.Unlock()
-
-	// Same NC+MAC already in NC state - this is a repeat CreateNC, not stale.
-	if service.ncAlreadyExistsForMAC(ncID, mac) {
-		return nil
-	}
-
-	containerID, ipInfo := service.findEndpointStateByMAC(mac)
-	if ipInfo == nil {
-		return nil
-	}
-
-	defer func() {
-		result := "success"
-		errMsg := ""
-		if returnErr != nil {
-			result = "failure"
-			errMsg = returnErr.Error()
-		}
-		logger.SendMetric(aitelemetry.Metric{ //nolint:staticcheck // TODO: migrate to zap logger
-			Name:  logger.StaleHNSCleanupMetricStr,
-			Value: 1.0,
-			CustomDimensions: map[string]string{
-				"ncID":          ncID,
-				"mac":           mac,
-				"containerID":   containerID,
-				"hnsEndpointID": ipInfo.HnsEndpointID,
-				"hnsNetworkID":  ipInfo.HnsNetworkID,
-				"result":        result,
-				"errorMsg":      errMsg,
-			},
-		})
-	}()
-
-	logger.Printf("[cleanupStaleHNSResources] cleaning up stale HNS resources for container %s (MAC %s)", containerID, mac) //nolint:staticcheck // TODO: migrate to zap logger
-	if ipInfo.HnsEndpointID != "" {
-		if err := service.hnsClient.DeleteEndpointByID(ipInfo.HnsEndpointID); err != nil {
-			return fmt.Errorf("failed to delete stale HNS endpoint %s: %w", ipInfo.HnsEndpointID, err)
-		}
-	}
-	if ipInfo.HnsNetworkID != "" {
-		if err := service.hnsClient.DeleteNetworkByID(ipInfo.HnsNetworkID); err != nil {
-			return fmt.Errorf("failed to delete stale HNS network %s: %w", ipInfo.HnsNetworkID, err)
-		}
-	}
-
-	err := service.DeleteEndpointStateHelper(containerID)
-	if err != nil {
-		return fmt.Errorf("failed to remove stale endpoint state for container %s: %w", containerID, err)
-	}
-
-	logger.Printf("[cleanupStaleHNSResources] Successfully removed stale endpoint state and HNS resources for container %s", containerID) //nolint:staticcheck // TODO: migrate to zap logger
-	return nil
-}
