+++ cns/restserver/ipam_linux.go
@@ -1,0 +1,5 @@
+package restserver
+
+func (service *HTTPRestService) cleanupStaleHNSResources(_, _ string) error {
+	return nil
+}
